+++ jest.setup.js
@@ -1,3 +1,0 @@
-import chrome from 'jest-chrome'
-
-Object.assign(global, chrome)